# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'FAMA API Documentation'
  description: 'The Fama mobile API (v1) — token-authenticated endpoints for talents, brands and admins, sharing the same services and JSON envelope as the web app.'
  version: 1.0.0
servers:
  - url: 'https://fama.test'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain a token from one of the <b>login</b> or <b>register</b> endpoints, then send it as <code>Authorization: Bearer {token}</code>. Tokens are ability-scoped to the entity (talent / brand / admin) they belong to.'
security:
  -
    default: []
paths:
...
  '/api/v1/brand/deals/{deal_id}':
    get:
      summary: 'Get a deal room'
      operationId: getADealRoom
      description: "The deal, its stepper, the message timeline and whether the brand can act\non the current step. Reading the room marks the talent's messages read."
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  data: null
                  message: Unauthenticated.
                  errors: null
                  meta: null
                properties:
                  success:
                    type: boolean
                    example: false
                  data:
                    type: string
                    example: null
                    nullable: true
                  message:
                    type: string
                    example: Unauthenticated.
                  errors:
                    type: string
                    example: null
                    nullable: true
                  meta:
                    type: string
                    example: null
                    nullable: true
      tags:
        - 'Brand · Deals'
    parameters:
      -
        in: path
        name: deal_id
        description: 'The ID of the deal.'
        example: 1
        required: true
        schema:
          type: integer
...